requirements.txt
requests
sqlglot
PyGithub
openai
google-genai
